src/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
websockets==12.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.14.0
apache-beam[aws]==2.50.0  # Updated to include AWS support
joblib==1.3.2  # Added for model persistence
optuna==3.5.0  # Added for hyperparameter tuning
asyncpg==0.29.0  # Added for PostgreSQL async support
boto3==1.34.0  # Added for AWS S3 integration
pyarrow==11.0.0  # Compatible with Apache Beam 2.50.0
pyyaml==6.0.1  # Added for YAML configuration
# Development dependencies for load testing
locust==2.17.0  # Added for load testing
websocket-client==1.7.0  # Added for WebSocket testing
aiohttp==3.9.1  # Added for async HTTP testing